# docs/openapi/spec/membershipfees.yaml
...
openapi: 3.1.0
info:
  title: Klabis API — Membership fees module
  description: The Membership fees slice of the Klabis API. See klabis.yaml for the full API.
  version: 0.1.0
servers:
  - url: https://localhost:8443
    description: Local development server
  - url: https://api.klabis.com
    description: Production server
paths:
...
  /api/membership-fee-tiers/{id}:
    get:
      operationId: getTier
      tags: [MembershipFeeTiers]
      summary: Get membership fee tier details
      description: Returns a single membership fee tier by ID.
      security:
        - KlabisAuth: [MEMBERS]
      # No x-klabis-authority: readable by any authenticated caller. editTier/deleteTier
      # affordances still require MEMBERS:MANAGE via klabisAfford's own authorization check.
      parameters:
        - $ref: '#/components/parameters/TierIdParam'
      responses:
        '200':
          description: Membership fee tier found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipFeeTierResponse'
            application/prs.hal-forms+json:
              schema:
                $ref: '#/components/schemas/EntityModelMembershipFeeTierResponse'
          x-hal-links:
            self:
              description: This tier
            collection:
              operation: listTiers
              description: Back to the tier list
            rules:
              operation: listRules
              description: Payment rules for this tier
          x-hal-templates:
            default:
              operation: editTier
              description: Present only for callers with MEMBERS:MANAGE
            deleteTier:
              operation: deleteTier
              description: Present only for callers with MEMBERS:MANAGE
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'

    patch:
      operationId: editTier
      tags: [MembershipFeeTiers]
      summary: Edit a membership fee tier
      description: |
        Partial update of a tier's name, yearly fee and/or payment rules. Fields omitted from the
        request are left unchanged.
      security:
        - KlabisAuth: [MEMBERS]
      x-klabis-authority: MEMBERS_MANAGE
      parameters:
        - $ref: '#/components/parameters/TierIdParam'
      requestBody:
        required: true
        x-codegen-request-body-name: request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditMembershipFeeTierRequest'
      responses:
        '204':
          description: Membership fee tier updated
          content:
            application/prs.hal-forms+json: {}
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'

    delete:
      operationId: deleteTier
      tags: [MembershipFeeTiers]
      summary: Delete a membership fee tier
      description: Deletes a membership fee tier from the catalog.
      security:
        - KlabisAuth: [MEMBERS]
      x-klabis-authority: MEMBERS_MANAGE
      parameters:
        - $ref: '#/components/parameters/TierIdParam'
      responses:
        '204':
          description: Membership fee tier deleted
          content:
            application/prs.hal-forms+json: {}
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'
...
  /api/membership-fee-groups/{id}:
    # getFeeGroup returns the plain MembershipFeeGroupResponse payload. The group-members collection is
    # declared by the controller via HalResponseContext.embed(...) and rendered under _embedded
    # by HalResponseBodyAdvice — the same deferred-assembly path as _links and _templates, so it
    # stays out of the generated return type — same as events.yaml's getEvent.
    # Note the _embedded key below is "members" per MembershipFeeGroupResponse's
    # @Relation(collectionRelation = "members") — NOT the class-name-derived default.
    get:
      operationId: getFeeGroup
      tags: [MembershipFeeGroups]
      summary: Get membership fee group details with snapshot and member count
      description: |
        Returns a published fee group's snapshot (yearly fee, payment rules) and its members,
        embedded under "members".
      security:
        - KlabisAuth: [MEMBERS]
      # No x-klabis-authority: readable by any authenticated caller. assignMember/editSnapshot
      # affordances still require MEMBERS:MANAGE via klabisAfford's own authorization check.
      parameters:
        - $ref: '#/components/parameters/GroupIdParam'
      responses:
        '200':
          description: Membership fee group found
          content:
            application/prs.hal-forms+json:
              schema:
                $ref: '#/components/schemas/EntityModelMembershipFeeGroupResponseWithMembers'
          x-hal-links:
            self:
              description: This fee group
            rules:
              operation: listGroupRules
              description: Payment rules snapshot for this group
            sourceLevel:
              operation: getTier
              description: The membership fee tier this group's snapshot was published from
          x-hal-templates:
            default:
              operation: editSnapshot
              description: Present for callers with MEMBERS:MANAGE, only while the group is EDITABLE
            assignMember:
              operation: assignMember
              description: Present only for callers with MEMBERS:MANAGE
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'

    patch:
      operationId: editSnapshot
      tags: [MembershipFeeGroups]
      summary: Edit yearly fee and payment rules of a published level
      description: |
        Updates the yearly fee amount and/or payment rules snapshot of a published fee group. Only
        allowed while the group's status is EDITABLE.
      security:
        - KlabisAuth: [MEMBERS]
      x-klabis-authority: MEMBERS_MANAGE
      parameters:
        - $ref: '#/components/parameters/GroupIdParam'
      requestBody:
        required: true
        x-codegen-request-body-name: request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditGroupSnapshotRequest'
      responses:
        '204':
          description: Fee group snapshot updated
          content:
            application/prs.hal-forms+json: {}
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'
...
  /api/members/{memberId}/fee-choice/{year}:
    get:
      operationId: getChoice
      tags: [MemberFeeChoice]
      summary: Get member's current fee level choice for a year
      description: |
        Returns the member's current fee group choice and the recommended tier for the year, if
        any.
      security:
        - KlabisAuth: [MEMBERS]
      # No x-klabis-authority: ONLY the member themselves, no MANAGE-authority alternative.
      # See module header comment.
      x-klabis-owner-visible: true
      x-spring-provide-args:
        - '@com.klabis.members.ActingMember com.klabis.members.MemberId actingMember'
      parameters:
        - $ref: '#/components/parameters/FeeChoiceMemberIdParam'
        - $ref: '#/components/parameters/FeeChoiceYearParam'
      responses:
        '200':
          description: Fee level choice retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberFeeChoiceResponse'
            application/prs.hal-forms+json:
              schema:
                $ref: '#/components/schemas/EntityModelMemberFeeChoiceResponse'
          x-hal-links:
            self:
              description: This fee choice
            currentGroup:
              operation: getFeeGroup
              description: Present when the member has a current fee group choice
            recommendedLevel:
              operation: getTier
              description: Present when a recommended tier exists for this year
          x-hal-templates:
            default:
              operation: chooseTier
              description: Choose a fee level for this year
            removeChoice:
              operation: removeChoice
              description: Remove the fee level choice for this year
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          description: Forbidden - members can only access their own fee level choice
          content:
            application/problem+json:
              schema:
                $ref: './_shared/problem.yaml#/components/schemas/ProblemDetail'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'

    post:
      operationId: chooseTier
      tags: [MemberFeeChoice]
      summary: Choose a fee level for a year
      description: Records the member's chosen fee group for the given year.
      security:
        - KlabisAuth: [MEMBERS]
      # See getChoice above — ONLY the member themselves.
      x-klabis-owner-visible: true
      x-spring-provide-args:
        - '@com.klabis.members.ActingMember com.klabis.members.MemberId actingMember'
      parameters:
        - $ref: '#/components/parameters/FeeChoiceMemberIdParam'
        - $ref: '#/components/parameters/FeeChoiceYearParam'
      requestBody:
        required: true
        x-codegen-request-body-name: request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChooseFeeChoiceRequest'
      responses:
        '204':
          description: Fee level choice recorded
          content:
            application/prs.hal-forms+json: {}
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          description: Forbidden - members can only manage their own fee level choice
          content:
            application/problem+json:
              schema:
                $ref: './_shared/problem.yaml#/components/schemas/ProblemDetail'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'

    delete:
      operationId: removeChoice
      tags: [MemberFeeChoice]
      summary: Remove fee level choice for a year
      description: Removes the member's chosen fee group for the given year.
      security:
        - KlabisAuth: [MEMBERS]
      # See getChoice above — ONLY the member themselves.
      x-klabis-owner-visible: true
      x-spring-provide-args:
        - '@com.klabis.members.ActingMember com.klabis.members.MemberId actingMember'
      parameters:
        - $ref: '#/components/parameters/FeeChoiceMemberIdParam'
        - $ref: '#/components/parameters/FeeChoiceYearParam'
      responses:
        '204':
          description: Fee level choice removed
          content:
            application/prs.hal-forms+json: {}
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          description: Forbidden - members can only manage their own fee level choice
          content:
            application/problem+json:
              schema:
                $ref: './_shared/problem.yaml#/components/schemas/ProblemDetail'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'
...
components:
  securitySchemes:
    KlabisAuth:
      type: oauth2
      description: OAuth2 authentication with Klabis Authorization Server
      flows:
        authorizationCode:
          authorizationUrl: /oauth2/authorize
          tokenUrl: /oauth2/token
          scopes:
            openid: authenticate
            profile: display user profile information
            MEMBERS: Members
            EVENTS: Events
  parameters:
    TierIdParam:
      name: id
      in: path
      description: Membership fee tier UUID
      required: true
      schema:
        type: string
        format: uuid
...
    GroupIdParam:
      name: id
      in: path
      description: Membership fee group UUID
      required: true
      schema:
        type: string
        format: uuid
...
    FeeChoiceMemberIdParam:
      name: memberId
      in: path
      description: Member UUID
      required: true
      # Marks the owner for operations that declare x-klabis-owner-visible. Inert on its own:
      # @OwnerId is only consulted for a method or field already marked @OwnerVisible, so the
      # operations here that do not opt into ownership are unaffected.
      x-klabis-owner-id: true
      schema:
        type: string
        format: uuid
    FeeChoiceYearParam:
      name: year
      in: path
      description: Calendar year
      required: true
      schema:
        type: integer
        format: int32
...
  schemas:
...
    EntityModelMembershipFeeTierResponse:
      allOf:
        - $ref: '#/components/schemas/MembershipFeeTierResponse'
        - type: object
          properties:
            _links:
              $ref: './_shared/hal.yaml#/components/schemas/Links'
            _templates:
              $ref: './_shared/hal.yaml#/components/schemas/HalFormsTemplates'
    MembershipFeeTierResponse:
      type: object
      description: Complete membership fee tier information.
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        yearlyFeeAmount:
          type: number
        yearlyFeeCurrency:
          type: string
...
    EditMembershipFeeTierRequest:
      type: object
      description: |
        Partial update. Fields omitted from the request are left unchanged. When rules is
        provided, it replaces the tier's entire rule set.
      properties:
        name:
          type: string
        yearlyFeeAmount:
          type: number
        yearlyFeeCurrency:
          type: string
        rules:
          type: array
          items:
            $ref: '#/components/schemas/PaymentRuleRequest'
    PaymentRuleRequest:
      type: object
      required: [eventTypeId, rankingShortName, ruleType]
      properties:
        eventTypeId:
          type: string
          format: uuid
        rankingShortName:
          type: string
          x-klabis-not-blank: true
        ruleType:
          # Plain string, not enum — see the CampaignStatusFilterParam comment above for why; same
          # generator limitation, and no existing domain RuleType enum to map onto (the domain uses
          # a sealed RuleValue interface, not an enum). The domain's switch expression already
          # rejects unknown values.
          type: string
          x-klabis-not-blank: true
        percent:
          type: integer
          format: int32
          description: Required when ruleType is PERCENTAGE.
        fixedAmount:
          type: number
          description: Required when ruleType is FIXED_AMOUNT.
        fixedCurrency:
          type: string
          description: Defaults to CZK when omitted.
...
    PaymentRuleResponse:
      type: object
      description: |
        A single payment rule. percentage is present when ruleType is PERCENTAGE; fixedAmount/
        fixedCurrency are present when ruleType is FIXED_AMOUNT.
      properties:
        eventTypeId:
          type: string
          format: uuid
        rankingShortName:
          type: string
        ruleType:
          type: string
          enum: [PERCENTAGE, FIXED_AMOUNT]
        percentage:
          type: integer
          format: int32
        fixedAmount:
          type: number
        fixedCurrency:
          type: string
...
    EntityModelMembershipFeeGroupResponseWithMembers:
      allOf:
        - $ref: '#/components/schemas/MembershipFeeGroupResponse'
        - type: object
          properties:
            _links:
              $ref: './_shared/hal.yaml#/components/schemas/Links'
            _templates:
              $ref: './_shared/hal.yaml#/components/schemas/HalFormsTemplates'
            _embedded:
              type: object
              properties:
                members:
                  type: array
                  items:
                    $ref: '#/components/schemas/MemberInGroupResponse'
    MembershipFeeGroupResponse:
      type: object
      description: A published fee group — a snapshot taken from a tier for one campaign year.
      properties:
        id:
          type: string
          format: uuid
        sourceLevelId:
          type: string
          format: uuid
        name:
          type: string
        year:
          type: integer
          format: int32
        yearlyFeeAmount:
          type: number
        yearlyFeeCurrency:
          type: string
        status:
          type: string
          enum: [EDITABLE, FROZEN]
        memberCount:
          type: integer
          format: int32
        rulesSnapshot:
          type: array
          items:
            $ref: '#/components/schemas/PaymentRuleResponse'
    MemberInGroupResponse:
      type: object
      description: A member's assignment to a fee group.
      x-klabis-relation:
        collectionRelation: members
      properties:
        memberId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        registrationNumber:
          type: string
        joinedAt:
          type: string
          format: date
        source:
          type: string
          enum: [MEMBER_CHOICE, ADMIN_ASSIGNMENT]
    EditGroupSnapshotRequest:
      type: object
      required: [yearlyFeeAmount]
      properties:
        yearlyFeeAmount:
          type: number
          exclusiveMinimum: 0
        yearlyFeeCurrency:
          type: string
          description: Defaults to CZK when omitted.
        rules:
          type: array
          items:
            $ref: '#/components/schemas/PaymentRuleRequest'
...
    EntityModelMemberFeeChoiceResponse:
      allOf:
        - $ref: '#/components/schemas/MemberFeeChoiceResponse'
        - type: object
          properties:
            _links:
              $ref: './_shared/hal.yaml#/components/schemas/Links'
            _templates:
              $ref: './_shared/hal.yaml#/components/schemas/HalFormsTemplates'
    MemberFeeChoiceResponse:
      type: object
      description: A member's fee level choice for a given year.
      properties:
        memberId:
          type: string
          format: uuid
        year:
          type: integer
          format: int32
        currentGroupId:
          type: string
          format: uuid
        recommendedLevelId:
          type: string
          format: uuid
    ChooseFeeChoiceRequest:
      type: object
      required: [membershipFeeGroupId]
      properties:
        membershipFeeGroupId:
          type: string
          format: uuid
...